refactor provisioning package
