+++ internal/controllers/provisioning/provisioners/pulumi/azure_db.go
@@ -71,7 +71,7 @@ func deployAzureDb(target provisioning.ProvisioningTarget,
 		ignoreChanges = []string{"resourceGroupName", "serverName", "elasticPoolId", "createMode", "sourceDatabaseId", "maxSizeBytes", "readScale", "requestedBackupStorageRedundancy", "catalogCollation", "collation", "sku", "zoneRedundant", "maintenanceConfigurationId", "isLedgerOn"}
 	}
 
-	dbNameV1 := provisioning.Match(target,
+	dbNameV1 := provisioning.MatchTarget(target,
 		func(tenant *platformv1.Tenant) string {
 			return fmt.Sprintf("%s_%s_%s", azureDb.Spec.DbName, tenant.Spec.PlatformRef, tenant.GetName())
 		},
